pyproject.toml
[project]
name = "intent-routed-agent-advanced"
version = "0.1.0"
description = "Production-grade multi-agent AI platform with MCP tool ecosystem"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-text-splitters>=0.0.1",
    "langgraph>=0.0.20",
    "langsmith>=0.0.87",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "tenacity>=8.2.3",
    "faiss-cpu>=1.7.4",
    "numpy>=1.26.0",
    "tiktoken>=0.5.2",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "langdetect>=1.0.9",
    "deep-translator>=1.11.4",
    "textblob>=0.17.1",
    "mcp>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "cli"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
